Fix for instance generator imports + added python test
