+++ python/src/ecole/instance_generators.py
@@ -1,0 +1,2 @@
+from ecole.setcover import *
+from ecole.combinatorial_auction import *
