--- python/src/ecole/instance_generators/instance_generators.py
@@ -1,2 +1,0 @@
-from setcover import Setcover 
-from combinatorial_auction import CombinatorialAuction
